+++ webnesta/WebApps/WebNesta.Coyote.Web/Models/DataComponentViewModel.cs
@@ -12,13 +12,13 @@ namespace WebNesta.Coyote.Web.Models
 
     public class Classe
     {
-        public int Id { get; set; }
+        public decimal Id { get; set; }
         public string Descricao { get; set; }
 
     }
     public class Modelo
     {
-        public int Id { get; set; }
+        public decimal Id { get; set; }
         public string Descricao { get; set; }
     }
 }
